web/requirements.txt
sanic
sqlalchemy
tortoise-orm
aiomysql
cerberus